Fixed black trails when custom bros throw grenades
The pocketted special grenades were being spawned by SpawnGrenadeOverNetwork, which has an RPC call to SetupGrenade where it passes in firedBy, which fails because RPC calls fail when passed custom bros / objects.

Fixing this actually causes another bug though, because SetupGrenade sets the firedBy variable, which GrenadeAirstrike uses when calling in missiles. So once it gets set to a custom bro, it will fail to call in missiles, because the call to SpawnProjectileOverNetwork will now pass in a custom bro rather than a null value like it was doing before. So we also have to patch SpawnProjectileOverNetwork.